+++ lib/database_helper/database_utils.dart
@@ -4,7 +4,11 @@ import 'model/symptoms_data.dart';
 
 List<Symptoms> defaultSymptomsData = [];
 
-var weekDayNameDayMonth = DateFormat('EEEE, d MMM'); /// Sunday, 10 Jun
-var defaultDateFormat = DateFormat('yyyy-MM-dd'); /// 2024-06-20
-var currentDateFormat =
-    DateFormat('yyyy-MM-dd hh:mm:ss'); /// 2024-06-20 12:20:11
+/// Sunday, 10 Jun
+var weekDayNameDayMonth = DateFormat('EEEE, d MMM');
+
+/// 2024-06-20
+var defaultDateFormat = DateFormat('yyyy-MM-dd');
+
+/// 2024-06-20 12:20:11
+var currentDateFormat = DateFormat('yyyy-MM-dd hh:mm:ss');