+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.1.5] - 19 June 2024
+* Added "viewType" of 'Menstrual Cycle Widget'
+
+## [0.1.4] - 18 June 2024
+* Added new params 
+* added functionality of store store user's symptoms 
+* added encryption functionality for storing user's data
+* Update README File description.
+
 ## [0.1.3] - 16 June 2024
 * Added new params
 * Update README File description.